+++ package.sh
@@ -1,1 +1,2 @@
 python3 setup.py sdist
+rm -r dist