+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG
 
 ## v0.10.1
 - refactored to eliminate lib folder
+- added on_end to Tween
+- realized IDLE is a good host
 
 ## v0.10.0
 - made a changelog